+++ examples/resources/illumio-core_provisioning/variables.tf
@@ -1,0 +1,4 @@
+variable "env_label_href" {
+  type        = string
+  description = "HREF of the label used as the rule set scope"
+}